Ports/bzip2: Set AR and RANLIB
bzip2's Makefile uses the '=' operator to set these variables so they
cannot be overridden by just the environment variables; we have to pass
them on the command line.

This change ensures that the system ar/ranlib are no longer used, so the
port can be build on macOS or non-x86 Linux.